fix: foldmarks: position offset due to logo
* This patch fixes the position offset of the foldmarks and the
recipient field of about 15mm
* The position offset was caused by an incorrect insertion of the logo
into the document header
* The logo has been made part of the actual document header `firsthead`
through the use of `tabularx` and `multirow` to construct a table in the
header. The logo is located in the left-most column while the personal
information is in the two right-most columns. The X-type column of
`tabularx` is used to stretch the table to the full width of the
document.